requirements.txt
requests
Mastodon.py
python-dotenv
hdfs
apache-airflow
mrjob
happybase